dump data and fix some bug

+++ backend/database/migrations/2023_10_16_073103_create_user_requests_table.php
@@ -19,7 +19,7 @@ class CreateUserRequestsTable extends Migration
             $table->string('type', 50);
             $table->string('status', 50);
             $table->timestamp('expired');
-            $table->integer('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->cascadeOnDelete()->cascadeOnUpdate();
             $table->timestamps();
         });
